Patch 1154
・コンパイルオプション /W4 有効時の警告修正
	-> warning C4189: 'xxx' : ローカル変数が初期化されましたが、参照されていません
	-> warning C4245: '=' : 'int' から 'unsigned int' に変換しました。signed/unsigned が一致しません。
	-> warning C4389: '!=' : signed と unsigned の数値を比較しようとしました。
	-> warning C4701: 初期化されていない可能性のあるローカル変数 'xxx' が使用されます
	-> warning C4702: 制御が渡らないコードです。
・イクシードブレイクの移動速度減少が、イクシードブレイクのスキルLvではなくフリーキャストのスキルLvを
  元に計算されていたのを修正（unit.c）
・ファイルのサイズを取得するとき、fseek, ftell を使わず stat を利用するように改善（httpd.c, grfio.c, map.c）
・map_mdmap_clean() のリファクタリング（map.c）

+++ src/common/socket.c
@@ -656,7 +656,7 @@ void realloc_fifo(int fd, size_t new_rfifo_size, size_t new_wfifo_size)
 
 	s = session[fd];
 	if (new_rfifo_size > 0 &&
-	    s->max_rdata - s->rdata != new_rfifo_size &&
+	    (size_t)(s->max_rdata  - s->rdata) != new_rfifo_size &&
 	    (size_t)(s->rdata_size - s->rdata) < new_rfifo_size)
 	{
 		unsigned char * p = s->rdata;
@@ -666,7 +666,7 @@ void realloc_fifo(int fd, size_t new_rfifo_size, size_t new_wfifo_size)
 		s->max_rdata  = s->rdata + new_rfifo_size;
 	}
 	if (new_wfifo_size > 0 &&
-	    s->max_wdata  - s->wdata != new_wfifo_size &&
+	    (size_t)(s->max_wdata  - s->wdata) != new_wfifo_size &&
 	    (size_t)(s->wdata_size - s->wdata) < new_wfifo_size)
 	{
 		unsigned char * p = s->wdata;
